refa: set config file for execution
